Move listLoan function to separate hook
